Extract some vehicle-related functions
